# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OrganizeAI",
        "version": "1.0"
    },
    "paths": {
        "/api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "List all documents",
                "operationId": "479e53af7d4820aecb8f9534fdbe21aa",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of documents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DocumentResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a new document",
                "operationId": "9b247d60e8834d470666fba8e773edb4",
                "requestBody": {
                    "description": "Document upload payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "document"
                                ],
                                "properties": {
                                    "document": {
                                        "description": "The document file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "Title of the document",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Tags associated with the document",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "metadata": {
                                        "description": "Additional metadata for the document",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a specific document",
                "operationId": "0fa5befcdb9042b346d01baea49dee95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of the document",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update a specific document",
                "operationId": "b851458865e322163fe9b88595919255",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Document update payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the document",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Tags associated with the document",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "metadata": {
                                        "description": "Additional metadata for the document",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a specific document",
                "operationId": "889c107c27dc6b118dee932d1760f574",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{id}/download": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Download a specific document",
                "operationId": "9c3894d9d7b13268e679345d76550898",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download URL generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "download_url": {
                                            "description": "URL to download the document",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DocumentResource": {
                "title": "Document",
                "required": [
                    "id",
                    "title",
                    "file_type"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the document",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "title": {
                        "description": "Title of the document",
                        "type": "string"
                    },
                    "file_type": {
                        "description": "Type of the file (e.g., pdf, jpg)",
                        "type": "string"
                    },
                    "metadata": {
                        "description": "Additional metadata related to the document",
                        "type": "object"
                    },
                    "tags": {
                        "description": "Tags associated with the document",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    },
                    "extractions": {
                        "description": "Extraction data from the document",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExtractionResource"
                        }
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ExtractionResource": {
                "title": "Extraction",
                "required": [
                    "id",
                    "content"
                ],
                "properties": {
                    "id": {
                        "description": "Extraction's unique identifier",
                        "type": "string"
                    },
                    "content": {
                        "description": "Extracted content from the document",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Extraction creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TagResource": {
                "title": "Tag",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Tag's unique identifier",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of the tag",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "User's unique identifier",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Documents",
            "description": "API Endpoints for Document Management"
        }
    ]
}